pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "3.0.0"
description = ""
authors = ["Timofey"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = ">=3.10,<4"
flask = "^2.2.3"
gunicorn = "^20.1.0"
flake8 = "^6.0.0"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
coverage = "^7.2.2"
flask-session = "^0.4.0"
python-dotenv = "^1.0.0"
bootstrap-flask = "^2.2.0"
flask-wtf = "^1.1.1"
psycopg2-binary = "^2.9.6"
validators = "^0.20.0"
requests = "^2.28.2"
beautifulsoup4 = "^4.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
